requirements.txt
streamlit==1.50.0
openai==2.0.0
python-dotenv==1.1.1
beautifulsoup4==4.14.2
requests==2.32.5
markdownify==1.2.0
selenium==4.35.0
webdriver-manager==4.0.2